refa(noop): rename gen module to grpc
